actual functions hotfix

+++ shared/src/androidMain/kotlin/com/ekdorn/classicdungeon/shared/dependant/GLFunctions.kt
@@ -37,6 +37,6 @@ actual object GLFunctions {
     ) {
     }
 
-    actual fun image(w: Int, h: Int, data: IntArray) {
+    actual fun image(w: Int, h: Int, data: ByteArray) {
     }
 }
