+++ shared/src/desktopMain/kotlin/com/ekdorn/classicdungeon/shared/dependant/GLFunctions.kt
@@ -38,6 +38,6 @@ actual object GLFunctions {
     ) {
     }
 
-    actual fun image(w: Int, h: Int, data: IntArray) {
+    actual fun image(w: Int, h: Int, data: ByteArray) {
     }
 }
